Extract JdbcDiagnosticModule
