+++ presto-base-jdbc/src/main/java/io/prestosql/plugin/jdbc/JdbcConnectorFactory.java
@@ -16,7 +16,6 @@ package io.prestosql.plugin.jdbc;
 import com.google.inject.Injector;
 import com.google.inject.Module;
 import io.airlift.bootstrap.Bootstrap;
-import io.prestosql.plugin.base.jmx.MBeanServerModule;
 import io.prestosql.plugin.jdbc.credential.CredentialProviderModule;
 import io.prestosql.spi.classloader.ThreadContextClassLoader;
 import io.prestosql.spi.connector.Connector;
@@ -24,7 +23,6 @@ import io.prestosql.spi.connector.ConnectorContext;
 import io.prestosql.spi.connector.ConnectorFactory;
 import io.prestosql.spi.connector.ConnectorHandleResolver;
 import io.prestosql.spi.type.TypeManager;
-import org.weakref.jmx.guice.MBeanModule;
 
 import java.util.Map;
 
@@ -69,8 +67,6 @@ public class JdbcConnectorFactory
                     binder -> binder.bind(TypeManager.class).toInstance(context.getTypeManager()),
                     new JdbcModule(catalogName),
                     new CredentialProviderModule(),
-                    new MBeanServerModule(),
-                    new MBeanModule(),
                     module);
 
             Injector injector = app
