+++ presto-base-jdbc/src/main/java/io/prestosql/plugin/jdbc/JdbcModule.java
@@ -16,24 +16,14 @@ package io.prestosql.plugin.jdbc;
 import com.google.inject.Binder;
 import com.google.inject.Key;
 import com.google.inject.Module;
-import com.google.inject.Provides;
 import com.google.inject.Scopes;
-import com.google.inject.Singleton;
-import io.airlift.log.Logger;
-import io.prestosql.plugin.base.util.LoggingInvocationHandler;
-import io.prestosql.plugin.base.util.LoggingInvocationHandler.ReflectiveParameterNamesProvider;
-import io.prestosql.plugin.jdbc.jmx.StatisticsAwareConnectionFactory;
-import io.prestosql.plugin.jdbc.jmx.StatisticsAwareJdbcClient;
 import io.prestosql.spi.connector.ConnectorAccessControl;
 import io.prestosql.spi.procedure.Procedure;
 
-import static com.google.common.reflect.Reflection.newProxy;
 import static com.google.inject.multibindings.Multibinder.newSetBinder;
 import static com.google.inject.multibindings.OptionalBinder.newOptionalBinder;
 import static io.airlift.configuration.ConfigBinder.configBinder;
-import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
-import static org.weakref.jmx.guice.ExportBinder.newExporter;
 
 public class JdbcModule
         implements Module
@@ -48,6 +38,8 @@ public class JdbcModule
     @Override
     public void configure(Binder binder)
     {
+        binder.install(new JdbcDiagnosticModule(catalogName));
+
         newOptionalBinder(binder, ConnectorAccessControl.class);
         newSetBinder(binder, Procedure.class);
         newSetBinder(binder, SessionPropertiesProvider.class);
@@ -62,45 +54,5 @@ public class JdbcModule
                 .to(Key.get(JdbcClient.class, StatsCollecting.class));
         binder.bind(ConnectionFactory.class)
                 .to(Key.get(ConnectionFactory.class, StatsCollecting.class));
-
-        newExporter(binder).export(Key.get(JdbcClient.class, StatsCollecting.class))
-                .as(generator -> generator.generatedNameOf(JdbcClient.class, catalogName));
-        newExporter(binder).export(Key.get(ConnectionFactory.class, StatsCollecting.class))
-                .as(generator -> generator.generatedNameOf(ConnectionFactory.class, catalogName));
-    }
-
-    @Provides
-    @Singleton
-    @StatsCollecting
-    public JdbcClient createJdbcClientWithStats(@ForBaseJdbc JdbcClient client)
-    {
-        StatisticsAwareJdbcClient statisticsAwareJdbcClient = new StatisticsAwareJdbcClient(client);
-
-        Logger logger = Logger.get(format("io.prestosql.plugin.jdbc.%s.jdbcclient", catalogName));
-
-        JdbcClient loggingInvocationsJdbcClient = newProxy(JdbcClient.class, new LoggingInvocationHandler(
-                statisticsAwareJdbcClient,
-                new ReflectiveParameterNamesProvider(),
-                logger::debug));
-
-        return new ForwardingJdbcClient()
-        {
-            @Override
-            protected JdbcClient getDelegate()
-            {
-                if (logger.isDebugEnabled()) {
-                    return loggingInvocationsJdbcClient;
-                }
-                return statisticsAwareJdbcClient;
-            }
-        };
-    }
-
-    @Provides
-    @Singleton
-    @StatsCollecting
-    public static ConnectionFactory createConnectionFactoryWithStats(@ForBaseJdbc ConnectionFactory connectionFactory)
-    {
-        return new StatisticsAwareConnectionFactory(connectionFactory);
     }
 }
